Added reflekto_lib.h and last letters need to be fixed
