+++ components/libraries/gfx/fonts/font10x16.c
@@ -114,12 +114,14 @@ const unsigned char Font10x16[] = {
 0x40, 0x60, 0x30, 0x10, 0x30, 0x60, 0x40, 0x60, 0x30, 0x10, 0x08, 0x0c, 0x06, 0x02, 0x06, 0x0c, 0x08, 0x0c, 0x06, 0x02, // ~
 0x00, 0x80, 0xc0, 0x60, 0x30, 0x30, 0x60, 0xc0, 0x80, 0x00, 0x0f, 0x0f, 0x0c, 0x0c, 0x0c, 0x0c, 0x0c, 0x0c, 0x0f, 0x0f};// 
 
+
+
 /**
  * @brief Font descriptor, defines basic features of 10x16 font.
  */
 const nrf_gfx_font_desc_t Font10x16_desc = 
 {
-    .ascii_offset = 32,
+    .ascii_offset = 32 ,
     .width = 10,
     .height = 16,
     .space_size = 3,
